+++ packages/core/src/index.ts
@@ -34,6 +34,7 @@ export * from "./photoCategories";
 export * from "./pod";
 export * from "./profitability";
 export * from "./rates";
+export * from "./receipt";
 export * from "./savedPlaces";
 export * from "./schemas";
 export * from "./search";